add spec with training hunter

+++ src/main/scala/com/github/wkicior/gymhunter/domain/training/Training.scala
@@ -2,7 +2,7 @@ package com.github.wkicior.gymhunter.domain.training
 
 import java.time.LocalDateTime
 
-case class Training(id: Long, slotsAvailable: Long, bookings_open_at: String, start_date: String) {
+case class Training(id: Long, slotsAvailable: Int, bookings_open_at: String, start_date: String) {
   def canBeBooked(): Boolean = {
     this.slotsAvailable > 0 //&& bookings_open_at.compareTo(LocalDateTime.now()) > 1
   }
